YoutubeVideoSummarizer/requirements.txt
crewai
youtube_transcript_api
streamlit
python-dotenv
pathlib
google-generativeai
pytube